--- a/mantis/cli/deskew.py
+++ b/mantis/cli/deskew.py
@@ -5,6 +5,10 @@ from typing import List
 
 import click
 
+# Needed for multiprocessing with GPUs
+# https://github.com/pytorch/pytorch/issues/40403#issuecomment-1422625325
+import torch
+
 from iohub.ngff import open_ome_zarr
 
 from mantis.analysis.AnalysisSettings import DeskewSettings
@@ -12,10 +16,6 @@ from mantis.analysis.deskew import deskew_data, get_deskewed_data_shape
 from mantis.cli import utils
 from mantis.cli.parsing import config_filepath, input_position_dirpaths, output_dirpath
 from mantis.cli.utils import yaml_to_model
-
-# Needed for multiprocessing with GPUs
-# https://github.com/pytorch/pytorch/issues/40403#issuecomment-1422625325
-import torch
 
 torch.multiprocessing.set_start_method('spawn', force=True)