--- a/mantis/cli/deskew.py
+++ b/mantis/cli/deskew.py
@@ -4,9 +4,6 @@ from pathlib import Path
 from typing import List
 
 import click
-
-# Needed for multiprocessing with GPUs
-# https://github.com/pytorch/pytorch/issues/40403#issuecomment-1422625325
 import torch
 
 from iohub.ngff import open_ome_zarr
@@ -17,6 +14,8 @@ from mantis.cli import utils
 from mantis.cli.parsing import config_filepath, input_position_dirpaths, output_dirpath
 from mantis.cli.utils import yaml_to_model
 
+# Needed for multiprocessing with GPUs
+# https://github.com/pytorch/pytorch/issues/40403#issuecomment-1422625325
 torch.multiprocessing.set_start_method('spawn', force=True)