--- a/mantis/cli/deskew.py
+++ b/mantis/cli/deskew.py
@@ -94,5 +94,6 @@ def deskew(
             **deskew_args,
         )
 
+
 if __name__ == "__main__":
     deskew()