--- a/pyproject.toml
+++ b/pyproject.toml
@@ -41,6 +41,7 @@ dependencies = [
   "largestinteriorrectangle",
   "antspyx",
   "pystackreg",
+
 ]