--- a/pyproject.toml
+++ b/pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
   "ndtiff>=2.0",
   "nidaqmx",
   "numpy",
+  "monai",
   "pandas~=2.1",
   "pycromanager==0.28.1",
   "pydantic",