--- a/pyproject.toml
+++ b/pyproject.toml
@@ -42,7 +42,6 @@ dependencies = [
   "largestinteriorrectangle",
   "antspyx",
   "pystackreg",
-
 ]