@ieivanov 2024-07-11
We just discussed that the CLI call executed on CPU, as before. `deskew_data` has an option to use GPUs but that's not implemented in the CLI and that's OK because we speed things up with slurm and CPU multiprocessing. The option to deskew data on GPU is most useful for live data reconstruction / visualization, which currently happens through scripts.